Refactor components/UI to account for time travel deletion
