+++ src/App.svelte
@@ -1,8 +1,8 @@
 <script>
-  import ViewsContainer from './views/ViewsContainer.svelte';
-  import StatesContainer from './states/StatesContainer.svelte';
-  import TimeTravelContainer from './timetravel/TimeTravelContainer.svelte';
   import svelte from 'svelte/compiler';
+  import ViewsContainer from './left-panel/views/ViewsContainer.svelte';
+  import StatesContainer from './right-panel/states/StatesContainer.svelte';
+  
 
 
   // globals
@@ -144,11 +144,8 @@
 
 
 <main>
-  <section>
-    <ViewsContainer />
-    <StatesContainer />
-  </section>
-  <TimeTravelContainer />
+  <ViewsContainer />
+  <StatesContainer />
 </main>
 
 <style>
@@ -158,19 +155,8 @@
 		margin: 0;
 		padding: 0;
 		display: grid;
-    grid-template-rows: minmax(0, 1fr) 40px;
+    grid-template-columns: minmax(0, 1fr) minmax(0, 1fr);
 	}
-
-	section {
-    grid-row: 1 / span 1;
-    background-color: blueviolet;
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-  }
-
-	TimeTravelContainer {
-    grid-row: 2 / span 1;
-  }
 
 	ViewsContainer {
     grid-column: 1 / span 1;
