+++ src/left-panel/views/ViewsContainer.svelte
