+++ src/left-panel/views/ViewsDisplay.svelte
