+++ src/left-panel/views/ViewsNavbar.svelte
