+++ src/right-panel/states/StatesDisplay.svelte
