+++ src/right-panel/states/StatesNavbar.svelte
